refactor(router): shard session state by ID
